Move EncodeSeparate to a separate file
