+++ lib/include/sse/schemes/tethys/tethys_store.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <sse/schemes/abstractio/awonvm_vector.hpp>
 #include <sse/schemes/abstractio/kv_serializer.hpp>
 #include <sse/schemes/tethys/details/tethys_allocator.hpp>
